Default enable background verifier
As the background verifier is no longer deleting the invalid records,
we can safely default enable it. This commit sets the default
verification interval to 7 days which means each replica is verified
to exist every 7 days. This allows for Praefect to start producing
logs and metrics on invalid metadata prior to us ultimately enabling
the deletion in the future.