Add configuration to enable deletions in background verifier
Background verifier doesn't delete the metadata records it found for
non-existent replicas due to a possible race with renames. This race
will mostly manifest with Geo in use as Geo performs a lot of renames.
Non-geo setups mostly perform renames for soft-deletions so it might
be reasonabl to enable the deletion logic in non-geo setups. This
commit adds a config option to enable the verifier to delete invalid
metadata records if the admin so chooses.

Changelog: added