Resolve replica path in background verifier tests
To test scenarios where the replica pointed to by a metadata record
does not exist, the verifier deletes replicas from Gitalys as part of
its test setup. It does so by pointing the calls to the relative path
and does not resolve the replica path. This causes the tests to fail
when Praefect generates unique replica paths. This commit resolves the
replica path in the tests so the deletions always target the correct
repositories. As the helpers are mostly designed with Gitaly in mind,
the GetReplicaPath helper currently expects a Gitaly config which it
will use to dial the service. In Praefect's context, we don't have such
a config available nor do we even have the address of the server. This
commit expands the helper to take a ClientConn in to use optionally as
the other helpers do.