Makefile: Upgrade bundled Git to v2.36.1
We have recently introduced a new bundled Git v2.36.0.gl1. Closely
before we wanted to start the rollout of that version to production
though a new patch release v2.36.1 was released by the Git project.

Upgrade Git to this new version. Note that we directly replace our
bundled Git v2.36.0.gl1. This is fine though given that it has not yet
been rolled out.

Changelog: changed