protolint: Enforce new style for enumerations
Our enumerations do not currently have any strict design guidelines.
This brings two issues with it:

    1. Enumeration values are scoped per-service. Consequentially, if
       two enumerations want to declare a value with the same name, the
       declarations will clash.

    2. It is not possible to discern the case where a caller hasn't set
       the value of an enumeration at all versus the case where the
       field was explicitly set to its zero-value.

The official Protobuf style guide thus says that enum values should
always be prefixed with the name of the enum itself, and that the zero
value should be called `UNSPECIFIED`. Let's enforce this style via
protolint so that any new enumerations will follow this recommendation.