doc: Don't mention ancient gitaly-proto project anymore
In ancient times, Gitaly's Protobuf definitions had been living in a
separate project. This is still mentioned in some of our documentation,
even though it doesn't really play much of a role nowadays anymore.

Stop mentioning this old project: it's a thing of the past, and I'd be
surprised if anybody still needs information about it.