updateref: Do not return `HookError`s for the reference-transaction hook
We're returning specific `HookError`s from our hook updater because
Rails cares about the standard output and error streams of custom hooks
so that it can decide whether to display any messages contained in
either of them in the web interface. So while it does make sense to
pass up the hook outputs for our normal hooks, we're also creating a
`HookError` for the reference-transaction hook. This doesn't make a
whole lot of sense though: we do not support custom hooks for this hook
type, and thus the standard streams shouldn't contain any errors that
might want to be displayed in the UI.

Refactor the code to return normal errors when the execution of the
reference-transaction hook fails. This prepares for a refactoring where
we convert the `HookError` type to a `CustomHookError` type.