updateref: Convert `HookError` to `CustomHookError`
We have a `HookError` type that gets returned whenever the execution of
a Git hook fails. In general though the informaton that this error holds
is not of any importance, but we only really care about special-handling
errors raised by custom hooks.

Refactor the error to a more specialized `CustomHookError` to make it
more useful.