doc: Update Protobuf verification docs
The 'Contributing' section of the Protobuf doc still references the old
`gitaly-proto` project and incorrectly suggests that the source files
generated by `protoc` may vary with Go version.

This removes the outdated and incorrect information and adds a new
section with details on the `make` targets for linting and validating
protos.