ruby: Fix diverging Git configuration in Go and Ruby
Right now, we have multiple sources of truth for how Git is configured.
In Go, most of the configuration is handled centrally by Gitaly itself,
includes mandatory configuration that must always be set. This config
supersedes and overrides any values in the global gitconfig that is
installed by both Omnibus and CNG. The goal here is that Gitaly becomes
the single source of truth for the Git configuration eventually and that
it starts ignoring any system- or global-level configuration that exists
in the filesystem. With this schema, we have a bunch of benefits:

    - We exactly know about the configuration that Git commands are in.

    - We don't need to rely on the repository's gitconfig anymore like
      we did in the past.

    - We have much tighter control over the execution environment and
      can also easily set up per-command configuration.

    - We can conditionally inject Git configuration based on the Git
      version that's currently running.

While we already live in this world for quite a while in our Go code
base, the Ruby sidecar is a different story. Here we don't use any of
the central configuration yet, but instead we still rely on the files
installed by both CNG and Omnibus. Which effectively means that we
cannot currently stop installing them, or otherwise some important
config entries that we do alread inject in Go might get lost.

Unfortunately, this has come to bite us just now where we're close to
removing the sidecar altogether: starting with Git v2.36.0, it prints a
warning when `core.fsyncObjectFiles` is set. This means that we're
forced to conditionally either inject this configuration when running
with an older Git version, or alternatively we must inject the new Git
configuration `core.fsync`. And while this is easy enough to do in Go
land, the sidecar can't do that at all because it is forced to rely on
the gitconfig files installed by distributions.

Let's fix this issue by injecting the Git configuration as set up by the
Git command factory in the Ruby sidecar via environment variables. This
ensures that Git commands spawned by it will share their configuration
with what Go uses, including `core.fsyncObjectFiles` or `core.fsync`
depending on the current Git version. Given that we now know that those
options are always set in the context of Ruby we can then strip away the
old setting that is causing warnings in both CNG and Omnibus.

Changelog: fixed