ruby: Remove unused ability to fetch from remote repo in `#with_branch`
The `#with_branch` function can optionally fetch from a remote repo in
case the requested branch or commit does not yet exist in the local one.
It only has a single caller left though, and that caller doesn't ever
set the `start_repository` parameter to anything. Consequentially, we
know that it's always going to be the local repo anyway, making the
functionality unused.

Remove the ability to fetch from a start repository and the now-unused
`#fetch_sha` function and stop injecting configuration for git-fetch(1)
into the sidecar.