ruby: Read `HEAD` via Rugged
The `#head_symbolic_ref` function is using git-symbolic-ref(1) to read
the repo's `HEAD` symbolic reference. We're slowly phasing out all
executions of the Git command line though, and we can easily implement
it via Rugged.

Convert the function to use Rugged to read the reference. This allows us
to get rid of the now-unused `#run_git` function, and furthermore we
don't have to inject configuration for git-symbolic-ref(1) into the
sidecar anymore.