ruby: Remove unused `start_branch_name` parameter from `#with_branch`
The `start_branch` parameter from `#with_branch` could be used to
specify which branch the new branch should have been created from in
case it didn't exist yet. The only caller that still sets this is
setting it to the same value as `branch_name` though.

Remove this parameter and instead consolidate it into `branch_name` to
simplify the code.