git: Sync internal namespaces with distros' settings
Repositories hosted by Gitaly have several internal reference namespaces
that shouldn't be modified by clients e.g. by pushing to them. And while
we have the infrastructure in place already to ensure that writes are
restricted, the list of internal namespace is not in sync with the list
of namespaces configured in both Omnibus and CNG. Most notably, we're
missing both the `refs/tmp/` and `refs/remotes/` namespaces.

Add the missing namespaces to ensure that we're not regressing our
safety guards in this area when we remove the gitconfig in our distros.

Changelog: changed