testhelper: Make injected random feature flags more flaky
We cannot really exhaustively test some of our feature flags because
they're hit in almost every code path. To still heuristically hit both
code paths with the feature flag turned on and off we're thus using a
random value instead.

This randomness is seeded with the current time in milliseconds. Which
effectively means that even if a test uses multiple different contexts,
the likelihood that both will see the same feature flag value is rather
high because it's likely they'll run in the same millisecond. This is
hiding flakiness that happens in case a test implicitly relies on the
fact that the feature flag value must be the same across subtests when
by accident we're using multiple different contexts.

Let's inject more randomness to cause such tests to be a lot more flaky
than they are right now by instead using nanosecond resolution for the
randomness' seed.