Remove unnecessary dependencies on the build stage
The testing stage doesn't actually depend on any of the build stage's
artifacts. It only uses the cache which the build stage writes. The
caches are only expired if the relevant files changes like go.sum or
the Gemfile.lock. These caches merely speed up the testing jobs when
they are present. As such, we can remove the build dependency from the
test jobs. Most of the time, the files that make up the cache key are
not changed and thus the both the build step and the test step end up
using the cache. This speeds up the general case by running the steps
in parallel and making test failures visible earlier.

In the rarer case where the cache keys are changed, the test steps will
end up performing some extra work that the cache avoids. This seems like
a fair trade-off to speed up the general case.

There are three exceptions to this though:

'test:coverage' fails without the cache in place as it fails with permission
errors creating the directory for the test coverage report.

'verify' fails with a timeout if the cache is not in place.

'gosec_sast' depends on some files from the cache being moved into the right
place.

As the above jobs fail without the cache, the dependencies are still left
in place for them.