gitaly-lfs-smudge: Add tests which exercise binary on a high-level
While we do have tests which verify that parts of the gitaly-lfs-smudge
command run as expected, we don't have any which test the final binary.
This makes it hard to test for some specific scenarios and doesn't give
us an easy way to verify that it works as expected.

Add such tests to improve our test coverage.