gitaly-lfs-smudge: Fix panic when log config is missing
Because gitaly-lfs-smudge is executed by Git instead of by us we can't
directly log to either stdout or stderr. Instead, we're currently forced
to log into a separate file. The location of that file is controlled by
an environment variable: if present we set up logging, if not we skip
over the configuration.

There's a bug here though: we return a `nil` closer in case we skip
configuring the log and don't initialize the logger at all. First, this
leads to a panic because we unconditionally try to close the closer. And
second, even if this worked, because we don't configure the logger we'd
now log to stdout directly. The result is that the logs would end up in
the final blob's contents as seen by Git.

Fix this bug by always initializing the logger. In case no file is
given, we ask it to simply write to `io.Discard`.