repository: Convert GetArchive to use new smudge configuration
The `GetArchive()` RPC can optionally convert LFS pointers to their
actual contents via the gitaly-lfs-smudge filter. The configuration of
this filter happens via a set of environment variables, which we have
replaced with a single structure in this commit series.

Convert `GetArchive()` to use the new `smudge.Config` to inject the
environment.